+++ src/.net/Tenjin.Autofac.Tests/ExtensionsTests/AutofacContainerPublisherExtensionsTests.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 using Tenjin.Autofac.Extensions;
 using Tenjin.Autofac.Tests.Constants;
 using Tenjin.Autofac.Tests.Utilities;
-using Tenjin.Interfaces.Messaging;
+using Tenjin.Interfaces.Messaging.PublishSubscriber;
 
 namespace Tenjin.Autofac.Tests.ExtensionsTests
 {
