+++ src/.net/Tenjin.Autofac.Tests/Implementations/Messaging/EnumPublishers.cs
@@ -1,6 +1,6 @@
 ﻿using Tenjin.Autofac.Tests.Enums;
-using Tenjin.Implementations.Messaging;
-using Tenjin.Interfaces.Messaging;
+using Tenjin.Implementations.Messaging.PublisherSubscriber;
+using Tenjin.Interfaces.Messaging.PublishSubscriber;
 
 namespace Tenjin.Autofac.Tests.Implementations.Messaging
 {
