+++ src/.net/Tenjin.Autofac/Implementations/Messaging/PublisherSubscriber/AutofacPublisherRegistry.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
-using Tenjin.Implementations.Messaging;
-using Tenjin.Interfaces.Messaging;
+using Tenjin.Implementations.Messaging.PublisherSubscriber;
+using Tenjin.Interfaces.Messaging.PublishSubscriber;
 
-namespace Tenjin.Autofac.Implementations.Messaging
+namespace Tenjin.Autofac.Implementations.Messaging.PublisherSubscriber
 {
     public class AutofacPublisherRegistry<TKey, TData> : PublisherRegistry<TKey, TData> where TKey : notnull
     {
