+++ src/.net/Tenjin.Autofac/Tenjin.Autofac.csproj
@@ -26,7 +26,7 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="6.3.0" />
-    <PackageReference Include="Tenjin" Version="1.5.0.59" />
+    <PackageReference Include="Tenjin" Version="1.6.0.62" />
   </ItemGroup>
 	
   <ItemGroup>
