+++ src/azure-pipelines.yml
@@ -25,7 +25,7 @@ variables:
   coreVersion: "6.x"
   major: 1
   minor: 2
-  revision: 0
+  revision: 1
   buildNumber: $[counter(variables['major'], 0)]
   ${{ if or(contains(variables['Build.SourceBranch'], 'feature/'), contains(variables['Build.SourceBranch'], 'bug-fix/'))  }}:
     releaseType: '-alpha'
